hash personal info

+++ dimmer.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import time
+import hashlib
 import numpy as np
 from PIL import Image
 from PyQt5 import uic
@@ -14,7 +15,6 @@ from window_manager import WindowManager
 from RNN import RNN
 
 
-LIST = ['dimmer', '파일']
 form_class = uic.loadUiType("ui/main.ui")[0]
 
 
@@ -29,20 +29,17 @@ class Main(QMainWindow, form_class):
         self.WM = WindowManager()
         self.pre_window = self.WM.get_fore_window()
         self.rnn = RNN()
-        
+        self.runState = False
+
         self.startButton.clicked.connect(self.btn_clk_start)
         self.startState = True
         self.trainButton.clicked.connect(self.btn_clk_train)
         self.helpButton.clicked.connect(self.btn_clk_help)
+        self.helpState = True
 
-    def minimize_window(self, event):
-        if (event.type() == QEvent.WindowStateChange and 
-                self.isMinimized()):
-            self.setWindowFlags(self.windowFlags() & ~Qt.Tool)
-            self.trayIcon.show()
-            return True
-        else:
-            return super(Main, self).minimize_window(event)
+        self.timer = QTimer(self)
+        self.timer.start(200)
+        self.timer.timeout.connect(self.run)
 
     def restore_window(self, reason):
         if reason == QSystemTrayIcon.DoubleClick:
@@ -50,69 +47,68 @@ class Main(QMainWindow, form_class):
             self.showNormal()
 
     def btn_clk_start(self):
-        if self.startState == True:
+        if self.startState:
             self.startState = False
             self.startButton.setText('Stop')
             
+            self.runState = True
             self.back = Background()
             self.back.show()
             self.trayIcon.setVisible(True)
             self.hide()
-
-            self.timer = QTimer(self)
-            self.timer.start(1000)
-            self.timer.timeout.connect(self.run)
-
-            self.timer2 = QTimer(self)
-            self.timer2.start(3600000)
-            self.timer2.timeout.connect(self.train)
         else:
             self.startState = True
             self.startButton.setText('Start')
+
+            self.runState = False
             self.back.close()
             self.timer.stop()
-            self.timer2.stop()
 
     def btn_clk_train(self):
-        self.trainButton.setEnabled(False)
-        try:
+        if os.path.isfile('params.pkl'):
             os.remove('params.pkl')
-        except:
-            pass
-        time.sleep(1)
+        time.sleep(2)
+        self.rnn = RNN()
         self.rnn.train()
         QMessageBox.information(self, "RNN", "train finished")
-        self.trainButton.setEnabled(True)
 
     def btn_clk_help(self):
-        QMessageBox.information(self, "Help", "wnsrud3611@gmail.com")
+        if self.helpState:
+            self.helpState = False
+            self.helpButton.setText('go to tray icon')
+            QMessageBox.information(self, "Help", "wnsrud3611@gmail.com")
+        else:
+            self.helpState = True
+            self.helpButton.setText('Help')
+            self.trayIcon.setVisible(True)
+            self.hide()
+
+    def get_hash(self, text):
+        hash = hashlib.md5()
+        hash.update(text.encode())
+        
+        return hash.hexdigest()
 
     def run(self):
         try:
             with open("data.txt", 'a') as f:
                 cur_window = self.WM.get_fore_window().split()[0]
-                if cur_window != self.pre_window and cur_window != 'dimmer':
+                if cur_window != self.pre_window:
                     self.pre_window = cur_window
-                    f.write(" "+cur_window)
-                    target_windows = list(self.rnn.test([cur_window], 5))
-                    target_windows.append(cur_window)
-                    print(target_windows)
-                    self.WM.set_window(self.WM.find_window('dimmer'))
-                    all_windows = list(self.WM.get_windows())
-                    for t_window in target_windows:
-                        for i, window in enumerate(all_windows):
-                            if t_window == window.split()[0] and t_window not in LIST:
-                                self.WM.set_window(self.WM.find_window(all_windows[i]))
+                    cur_window_hash = self.get_hash(cur_window)
+                    f.write(" "+cur_window_hash)
+                    if self.runState:
+                        target_windows = list(self.rnn.test([cur_window_hash], 3))
+                        target_windows.append(cur_window_hash)
+                        print(target_windows)
+                        self.WM.set_window(self.WM.find_window('dimmer'))
+                        all_windows = list(self.WM.get_windows())
+                        for t_window in target_windows:
+                            for i, window in enumerate(all_windows):
+                                if t_window == self.get_hash(window.split()[0]):
+                                    self.WM.set_window(self.WM.find_window(all_windows[i]))
         except:
             pass
-
-    def train(self):
-        try:
-            os.remove('params.pkl')
-        except:
-            pass
-        time.sleep(1)
-        self.rnn.train()
 
 
 class Background(QMainWindow):
@@ -136,7 +132,6 @@ class Background(QMainWindow):
         self.setAttribute(Qt.WA_TranslucentBackground)
         self.setAttribute(Qt.WA_TransparentForMouseEvents)
         self.setWindowFlags(Qt.SplashScreen)
-
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
